+++ lib/src/ncontroller.dart
@@ -3,6 +3,7 @@ import 'package:flutter/foundation.dart';
 /// Base controller
 abstract class NController<T> {
   final List<void Function()> _listeners = [];
+  bool _disposed = false;
 
   /// Subscribe to changes with [listener]
   void addListener(void Function() listener) {
@@ -15,16 +16,28 @@ abstract class NController<T> {
   /// Update all listeners
   @protected
   void update() {
+    if (_disposed){
+      _disposedWarning('Can\'t update disposed controller');
+      return;
+    }
+
     final snapshot = List<void Function()>.from(_listeners);
     for (final l in snapshot) {
       try {
         l();
-      } catch (_) {}
+      } catch (ex, stack) {
+        debugPrint('NController error: $ex\n$stack');
+      }
     }
   }
 
-  /// First start [matate] and then update listeners
+  /// First start [mutate] and then update listeners
   void setAndUpdate(void Function() mutate) {
+    if (_disposed){
+      _disposedWarning('Can\'t update disposed controller');
+      return;
+    }
+
     mutate();
     update();
   }
@@ -37,11 +50,23 @@ abstract class NController<T> {
   @mustCallSuper
   void onReady() {}
 
-  /// Call on widget unmount
+  /// Call on widget dispose
   @mustCallSuper
-  void unmount() {}
+  void dispose() {
+    if (_disposed){
+      _disposedWarning('NController already disposed');
+      return;
+    }
+
+    _disposed = true;
+    _listeners.clear();
+  }
 
   /// Call before widget mount, its called before [onInit]
   @mustCallSuper
   void beforeMount() {}
+
+  /// Internal method to print warning when try to call disposed controller
+  void _disposedWarning(String message) =>
+      debugPrint('❌ Try to call disposed NController with type: ${runtimeType.toString()}\n$message');
 }